requirements.txt
Django==2.1
django-constance[database]==2.2.0

djangorestframework==3.8.2
django-filter==2.0.0

celery==4.2.1
redis==2.10.6
requests==2.19.1